refactor(setup): extract shared prompt symbols module
The two GitHub prompt files duplicated the unicode/ASCII symbol table
and the `supportsUnicode` / `symbol` helpers verbatim. Pull them into
`prompt-symbols.ts`; each prompt keeps only the symbols unique to it
(`ACTION_ARROW`, `CHECKBOX_PARTIAL`).
